more stuff

+++ src/func_80017C3C.c
@@ -3,7 +3,7 @@
 extern s16 D_8005E460;
 extern s16 D_8005E468;
 
-void func_80017C3C(s16 arg0) {
+void func_80017C3C(s32 arg0) {
     D_8005E468 = arg0;
     D_8005E460 = arg0;
 }
